2020.1.9

+++ shundi-privider-8001/src/main/java/com/pps/service/impl/OrderServiceImpl.java
@@ -3,12 +3,17 @@ package com.pps.service.impl;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import com.mongodb.WriteResult;
 import com.pps.mapper.TbOrderMapper;
 import com.pps.pojo.TbOrder;
 import com.pps.pojo.TbOrderExample;
 import com.pps.pojo.exception.UnknowException;
+import com.pps.pojo.group.Result;
 import com.pps.pojo.mongo.UserPoint;
+import com.pps.pojo.status.OrderStatus;
+import com.pps.pojo.status.PackageStatus;
 import com.pps.service.OrderService;
+import com.pps.service.UserPointService;
 import jdk.nashorn.internal.runtime.options.Option;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -18,6 +23,8 @@ import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
 
@@ -36,24 +43,29 @@ public class OrderServiceImpl implements OrderService {
     TbOrderMapper tbOrderMapper;
     @Autowired
     MongoTemplate mongoTemplate;
+    @Autowired
+    UserPointService userPointService;
 
 
     @Override
     public TbOrder addOrder(TbOrder tbOrder) {
 
+        UserPoint userPoint = userPointService.getUserPoint(tbOrder.getUserpointid());
+        int status=userPoint.getStatus();
+        if(status!=PackageStatus.已提交.getCode()){
+            throw new UnknowException("该货物此状态不可接单 请刷新页面 信息存在滞后");
+        }
+
+        tbOrder.setStatus(OrderStatus.已申请.getCode());
         tbOrderMapper.insert(tbOrder);
         Optional.of(tbOrder.getId()).orElseThrow(UnknowException::new);
-        //更新物品状态
-    /*    UserPoint userPoint = mongoTemplate.findById(tbOrder.getUserpointid(), UserPoint.class, "userPoint");
-        userPoint.setStatus(0);
-        mongoTemplate.save(userPoint);*/
 
         Query query = new Query();
-        query.addCriteria(Criteria.where("_id").is(tbOrder.getUserpointid()));  //_id区分引号 "1"和1
-        Update update = Update.update("status", "0");
+        query.addCriteria(Criteria.where("_id").is(userPoint.get_id()));  //_id区分引号 "1"和1
+        Update update = Update.update("status", PackageStatus.已申请.getCode());
         mongoTemplate.updateFirst(query,update,"userPoint");
-
         return  tbOrder;
+
     }
 
     @Override
@@ -61,20 +73,37 @@ public class OrderServiceImpl implements OrderService {
 
         int i = tbOrderMapper.updateByPrimaryKeySelective(tbOrder);
 
-        Boolean s = Optional.of(i == 1).orElseThrow(() -> {
-            return new UnknowException("更新订单失败");
-        });
+        if(i >1) {
+            throw  new UnknowException("更新订单失败");
+        };
 
-          return  Optional.of(s);
+          return  Optional.of(true);
     }
 
     @Override
     public Boolean deleteOrder(TbOrder tbOrder) {
-      int c=  tbOrderMapper.deleteByPrimaryKey(tbOrder.getId());
-      if(c>1){
-          throw new UnknowException("删除失败");
-      }
-      return  Optional.of(c==1).get();
+
+        int status = tbOrder.getStatus();
+        if(status==OrderStatus.已申请.getCode()||status==OrderStatus.被拒绝.getCode()){
+
+            if(status==OrderStatus.已申请.getCode()){
+
+                Query query = new Query();
+                query.addCriteria(Criteria.where("_id").is(tbOrder.getUserpointid()));  //_id区分引号 "1"和1
+                Update update = Update.update("status", PackageStatus.已提交.getCode());
+                WriteResult userPoint1 = mongoTemplate.updateFirst(query, update, "userPoint");
+
+
+            }
+            int c=  tbOrderMapper.deleteByPrimaryKey(tbOrder.getId());
+            if(c>1){
+                throw new UnknowException("删除失败");
+            }
+            return  Optional.of(c==1).get();
+
+        }else{
+            throw  new UnknowException("该状态下不可以改变其状态");
+        }
     }
 
     @Override
@@ -86,7 +115,12 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public PageInfo findByCondition(TbOrder tbOrder,Integer pageNum,Integer pageSize) {
-
+         if(pageNum==null) {
+             pageNum=1;
+         }
+         if(pageSize==null){
+             pageSize=10;
+         }
         TbOrderExample tbOrderExample=new TbOrderExample();
         TbOrderExample.Criteria criteria = tbOrderExample.createCriteria();
 
@@ -114,5 +148,28 @@ public class OrderServiceImpl implements OrderService {
         PageInfo pageInfo = new PageInfo<>(tbOrders);
 
         return pageInfo;
+    }
+
+    @Override
+    public Result getOrderNum(Integer userId) {
+
+        HashMap hashMap=new HashMap();
+        OrderStatus[] values = OrderStatus.values();
+        Arrays.stream(values).forEach(p->{
+
+            Integer code = p.getCode();
+            TbOrderExample tbOrderExample=new TbOrderExample();
+            TbOrderExample.Criteria criteria = tbOrderExample.createCriteria();
+            criteria.andStatusEqualTo(code);
+            criteria.andUseridEqualTo(userId);
+            int count = tbOrderMapper.countByExample(tbOrderExample);
+            hashMap.put(code+"",count);
+        });
+
+        return  new Result(true,hashMap);
+
+
+
+
     }
 }
